+++ ui/main/src/app/services/admin/CrudEntitiesService.ts
@@ -8,8 +8,8 @@
  */
 
 import {Observable, map} from 'rxjs';
-import {EntitiesService} from '../../../services/entities/EntitiesService';
-import {CrudUtilities} from './crudUtils';
+import {EntitiesService} from '../entities/EntitiesService';
+import {CrudUtilities} from './CrudUtils';
 
 export class CrudEntitiesService {
     getAll(): Observable<Array<any>> {
